feat(EntityRegistry): Add replicate functionality for suppliers
Introduces the ability to replicate a supplier entity as a client. This is achieved by adding an `onReplicate` prop to `EntityRegistry` and integrating a new state `addToClients` within the component. When a supplier is saved, if the replicate option is active, the supplier data is also used to create a new client entity.